feat: only remove old icons when regenerating all icons

+++ icons/build-favicons.sh
@@ -30,7 +30,9 @@ HOMELAB_ENV="$mode"
 
 input_dir="$(git rev-parse --show-toplevel)/icons"
 output_dir="$(git rev-parse --show-toplevel)/docker-images/custom/favicons/icons"
-rm -rf "$output_dir"
+if [ "${only_pattern:-}" = '' ]; then
+    rm -rf "$output_dir"
+fi
 mkdir -p "$output_dir"
 
 tmpdir=''
